+++ sdk/src/transactions/tool.rs
@@ -1,8 +1,11 @@
-use crate::{
-    idents::{move_std, primitives, pure_arg, sui_framework, workflow},
-    sui,
-    types::{NexusObjects, ToolMeta},
-    ToolFqn,
+use {
+    crate::{
+        idents::{move_std, pure_arg, sui_framework, workflow},
+        sui,
+        types::{NexusObjects, ToolMeta},
+        ToolFqn,
+    },
+    std::vec,
 };
 
 /// PTB template for registering a new Nexus Tool.
@@ -13,7 +16,7 @@ pub fn register_off_chain_for_self(
     address: sui::types::Address,
     collateral_coin: &sui::types::ObjectReference,
     invocation_cost: u64,
-) -> anyhow::Result<sui::types::Argument> {
+) -> anyhow::Result<()> {
     // `self: &mut ToolRegistry`
     let tool_registry = tx.input(sui::tx::Input::shared(
         *objects.tool_registry.object_id(),
@@ -51,7 +54,7 @@ pub fn register_off_chain_for_self(
     ));
 
     // `nexus_workflow::tool_registry::register_off_chain_tool()`
-    let owner_cap_over_tool = tx.move_call(
+    let result = tx.move_call(
         sui::tx::Function::new(
             objects.workflow_pkg_id,
             workflow::ToolRegistry::REGISTER_OFF_CHAIN_TOOL.module,
@@ -70,6 +73,20 @@ pub fn register_off_chain_for_self(
         ],
     );
 
+    // `tool: Tool`
+    let Some(tool) = result.nested(0) else {
+        return Err(anyhow::anyhow!(
+            "Failed to extract Tool from register_off_chain_tool result"
+        ));
+    };
+
+    // `owner_cap_over_tool: CloneableOwnerCap<OverTool>`
+    let Some(owner_cap_over_tool) = result.nested(1) else {
+        return Err(anyhow::anyhow!(
+            "Failed to extract OwnerCap<OverTool> from register_off_chain_tool result"
+        ));
+    };
+
     // `nexus_workflow::gas::deescalate()`
     let owner_cap_over_gas = tx.move_call(
         sui::tx::Function::new(
@@ -78,7 +95,7 @@ pub fn register_off_chain_for_self(
             workflow::Gas::DEESCALATE.name,
             vec![],
         ),
-        vec![tool_registry, owner_cap_over_tool, fqn],
+        vec![tool, owner_cap_over_tool],
     );
 
     // `gas_service: &mut GasService`
@@ -90,6 +107,7 @@ pub fn register_off_chain_for_self(
 
     // `single_invocation_cost_mist: u64`
     let single_invocation_cost_mist = tx.input(pure_arg(&invocation_cost)?);
+
     // `nexus_workflow::gas::set_single_invocation_cost_mist`
     tx.move_call(
         sui::tx::Function::new(
@@ -100,67 +118,33 @@ pub fn register_off_chain_for_self(
         ),
         vec![
             gas_service,
-            tool_registry,
+            tool,
             owner_cap_over_gas,
-            fqn,
             single_invocation_cost_mist,
         ],
     );
 
-    // `CloneableOwnerCap<OverGas>`
-    let over_gas_type = sui::types::TypeTag::Struct(Box::new(sui::types::StructTag::new(
-        objects.primitives_pkg_id,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.module,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.name,
-        vec![sui::types::TypeTag::Struct(Box::new(
-            sui::types::StructTag::new(
-                objects.workflow_pkg_id,
-                workflow::Gas::OVER_GAS.module,
-                workflow::Gas::OVER_GAS.name,
-                vec![],
-            ),
-        ))],
-    )));
+    // `Tool`
+    let tool_type = workflow::into_type_tag(objects.workflow_pkg_id, workflow::ToolRegistry::TOOL);
 
-    // `CloneableOwnerCap<OverTool>`
-    let over_tool_type = sui::types::TypeTag::Struct(Box::new(sui::types::StructTag::new(
-        objects.primitives_pkg_id,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.module,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.name,
-        vec![sui::types::TypeTag::Struct(Box::new(
-            sui::types::StructTag::new(
-                objects.workflow_pkg_id,
-                workflow::ToolRegistry::OVER_TOOL.module,
-                workflow::ToolRegistry::OVER_TOOL.name,
-                vec![],
-            ),
-        ))],
-    )));
+    // `sui::transfer::share_object`
+    tx.move_call(
+        sui::tx::Function::new(
+            sui_framework::PACKAGE_ID,
+            sui_framework::Transfer::PUBLIC_SHARE_OBJECT.module,
+            sui_framework::Transfer::PUBLIC_SHARE_OBJECT.name,
+            vec![tool_type],
+        ),
+        vec![tool],
+    );
 
     // `recipient: address`
     let recipient = sui_framework::Address::address_from_type(tx, address)?;
 
     // `sui::transfer::public_transfer`
-    tx.move_call(
-        sui::tx::Function::new(
-            sui_framework::PACKAGE_ID,
-            sui_framework::Transfer::PUBLIC_TRANSFER.module,
-            sui_framework::Transfer::PUBLIC_TRANSFER.name,
-            vec![over_tool_type],
-        ),
-        vec![owner_cap_over_tool, recipient],
-    );
+    tx.transfer_objects(vec![owner_cap_over_tool, owner_cap_over_gas], recipient);
 
-    // `sui::transfer::public_transfer`
-    Ok(tx.move_call(
-        sui::tx::Function::new(
-            sui_framework::PACKAGE_ID,
-            sui_framework::Transfer::PUBLIC_TRANSFER.module,
-            sui_framework::Transfer::PUBLIC_TRANSFER.name,
-            vec![over_gas_type],
-        ),
-        vec![owner_cap_over_gas, recipient],
-    ))
+    Ok(())
 }
 
 /// PTB template for registering a new onchain Nexus Tool.
@@ -177,7 +161,7 @@ pub fn register_on_chain_for_self(
     witness_id: sui::types::Address,
     collateral_coin: &sui::types::ObjectReference,
     address: sui::types::Address,
-) -> anyhow::Result<sui::types::Argument> {
+) -> anyhow::Result<()> {
     // `self: &mut ToolRegistry`
     let tool_registry = tx.input(sui::tx::Input::shared(
         *objects.tool_registry.object_id(),
@@ -221,7 +205,7 @@ pub fn register_on_chain_for_self(
     ));
 
     // `nexus_workflow::tool_registry::register_on_chain_tool()`
-    let owner_cap_over_tool = tx.move_call(
+    let result = tx.move_call(
         sui::tx::Function::new(
             objects.workflow_pkg_id,
             workflow::ToolRegistry::REGISTER_ON_CHAIN_TOOL.module,
@@ -242,41 +226,48 @@ pub fn register_on_chain_for_self(
         ],
     );
 
-    // `CloneableOwnerCap<OverTool>`
-    let over_tool_type = sui::types::TypeTag::Struct(Box::new(sui::types::StructTag::new(
-        objects.primitives_pkg_id,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.module,
-        primitives::OwnerCap::CLONEABLE_OWNER_CAP.name,
-        vec![sui::types::TypeTag::Struct(Box::new(
-            sui::types::StructTag::new(
-                objects.workflow_pkg_id,
-                workflow::ToolRegistry::OVER_TOOL.module,
-                workflow::ToolRegistry::OVER_TOOL.name,
-                vec![],
-            ),
-        ))],
-    )));
+    // `tool: Tool`
+    let Some(tool) = result.nested(0) else {
+        return Err(anyhow::anyhow!(
+            "Failed to extract Tool from register_off_chain_tool result"
+        ));
+    };
+
+    // `owner_cap_over_tool: CloneableOwnerCap<OverTool>`
+    let Some(owner_cap_over_tool) = result.nested(1) else {
+        return Err(anyhow::anyhow!(
+            "Failed to extract OwnerCap<OverTool> from register_off_chain_tool result"
+        ));
+    };
+
+    // `Tool`
+    let tool_type = workflow::into_type_tag(objects.workflow_pkg_id, workflow::ToolRegistry::TOOL);
+
+    // `sui::transfer::share_object`
+    tx.move_call(
+        sui::tx::Function::new(
+            sui_framework::PACKAGE_ID,
+            sui_framework::Transfer::PUBLIC_SHARE_OBJECT.module,
+            sui_framework::Transfer::PUBLIC_SHARE_OBJECT.name,
+            vec![tool_type],
+        ),
+        vec![tool],
+    );
 
     // `recipient: address`
     let recipient = sui_framework::Address::address_from_type(tx, address)?;
 
     // `sui::transfer::public_transfer`
-    Ok(tx.move_call(
-        sui::tx::Function::new(
-            sui_framework::PACKAGE_ID,
-            sui_framework::Transfer::PUBLIC_TRANSFER.module,
-            sui_framework::Transfer::PUBLIC_TRANSFER.name,
-            vec![over_tool_type],
-        ),
-        vec![owner_cap_over_tool, recipient],
-    ))
+    tx.transfer_objects(vec![owner_cap_over_tool], recipient);
+
+    Ok(())
 }
 
 /// PTB template for setting the invocation cost of a Nexus Tool.
 pub fn set_invocation_cost(
     tx: &mut sui::tx::TransactionBuilder,
     objects: &NexusObjects,
-    tool_fqn: &ToolFqn,
+    tool: &sui::types::ObjectReference,
     owner_cap: &sui::types::ObjectReference,
     invocation_cost: u64,
 ) -> anyhow::Result<sui::types::Argument> {
@@ -287,11 +278,11 @@ pub fn set_invocation_cost(
         true,
     ));
 
-    // `tool_registry: &mut ToolRegistry`
-    let tool_registry = tx.input(sui::tx::Input::shared(
-        *objects.tool_registry.object_id(),
-        objects.tool_registry.version(),
-        true,
+    // `tool: &Tool`
+    let tool = tx.input(sui::tx::Input::shared(
+        *tool.object_id(),
+        tool.version(),
+        false,
     ));
 
     // `owner_cap: &CloneableOwnerCap<OverGas>`
@@ -300,9 +291,6 @@ pub fn set_invocation_cost(
         owner_cap.version(),
         *owner_cap.digest(),
     ));
-
-    // `fqn: AsciiString`
-    let fqn = move_std::Ascii::ascii_string_from_str(tx, tool_fqn.to_string())?;
 
     // `single_invocation_cost_mist: u64`
     let single_invocation_cost_mist = tx.input(pure_arg(&invocation_cost)?);
@@ -315,13 +303,7 @@ pub fn set_invocation_cost(
             workflow::Gas::SET_SINGLE_INVOCATION_COST_MIST.name,
             vec![],
         ),
-        vec![
-            gas_service,
-            tool_registry,
-            owner_cap,
-            fqn,
-            single_invocation_cost_mist,
-        ],
+        vec![gas_service, tool, owner_cap, single_invocation_cost_mist],
     ))
 }
 
@@ -329,18 +311,15 @@ pub fn set_invocation_cost(
 pub fn unregister(
     tx: &mut sui::tx::TransactionBuilder,
     objects: &NexusObjects,
-    tool_fqn: &ToolFqn,
+    tool: &sui::types::ObjectReference,
     owner_cap: &sui::types::ObjectReference,
 ) -> anyhow::Result<sui::types::Argument> {
-    // `self: &mut ToolRegistry`
-    let tool_registry = tx.input(sui::tx::Input::shared(
-        *objects.tool_registry.object_id(),
-        objects.tool_registry.version(),
+    // `self: &mut Tool`
+    let tool = tx.input(sui::tx::Input::shared(
+        *tool.object_id(),
+        tool.version(),
         true,
     ));
-
-    // `fqn: AsciiString`
-    let fqn = move_std::Ascii::ascii_string_from_str(tx, tool_fqn.to_string())?;
 
     // `owner_cap: &CloneableOwnerCap<OverTool>`
     let owner_cap = tx.input(sui::tx::Input::owned(
@@ -364,7 +343,7 @@ pub fn unregister(
             workflow::ToolRegistry::UNREGISTER_TOOL.name,
             vec![],
         ),
-        vec![tool_registry, owner_cap, fqn, clock],
+        vec![tool, owner_cap, clock],
     ))
 }
 
@@ -373,13 +352,13 @@ pub fn unregister(
 pub fn claim_collateral_for_self(
     tx: &mut sui::tx::TransactionBuilder,
     objects: &NexusObjects,
-    tool_fqn: &ToolFqn,
+    tool: &sui::types::ObjectReference,
     owner_cap: &sui::types::ObjectReference,
 ) -> anyhow::Result<sui::types::Argument> {
-    // `self: &mut ToolRegistry`
-    let tool_registry = tx.input(sui::tx::Input::shared(
-        *objects.tool_registry.object_id(),
-        objects.tool_registry.version(),
+    // `self: &mut Tool`
+    let tool = tx.input(sui::tx::Input::shared(
+        *tool.object_id(),
+        tool.version(),
         true,
     ));
 
@@ -389,9 +368,6 @@ pub fn claim_collateral_for_self(
         owner_cap.version(),
         *owner_cap.digest(),
     ));
-
-    // `fqn: AsciiString`
-    let fqn = move_std::Ascii::ascii_string_from_str(tx, tool_fqn.to_string())?;
 
     // `clock: &Clock`
     let clock = tx.input(sui::tx::Input::shared(
@@ -408,7 +384,7 @@ pub fn claim_collateral_for_self(
             workflow::ToolRegistry::CLAIM_COLLATERAL_FOR_SELF.name,
             vec![],
         ),
-        vec![tool_registry, owner_cap, fqn, clock],
+        vec![tool, owner_cap, clock],
     ))
 }
 
@@ -526,11 +502,11 @@ mod tests {
     #[test]
     fn test_unregister_tool() {
         let objects = sui_mocks::mock_nexus_objects();
-        let tool_fqn = fqn!("xyz.dummy.tool@1");
+        let tool = sui_mocks::mock_sui_object_ref();
         let owner_cap = sui_mocks::mock_sui_object_ref();
 
         let mut tx = sui::tx::TransactionBuilder::new();
-        unregister(&mut tx, &objects, &tool_fqn, &owner_cap)
+        unregister(&mut tx, &objects, &tool, &owner_cap)
             .expect("Failed to build PTB for unregistering a tool.");
         let tx = sui_mocks::mock_finish_transaction(tx);
         let sui::types::TransactionKind::ProgrammableTransaction(
@@ -553,11 +529,11 @@ mod tests {
     #[test]
     fn test_claim_collateral_for_self() {
         let objects = sui_mocks::mock_nexus_objects();
-        let tool_fqn = fqn!("xyz.dummy.tool@1");
+        let tool = sui_mocks::mock_sui_object_ref();
         let owner_cap = sui_mocks::mock_sui_object_ref();
 
         let mut tx = sui::tx::TransactionBuilder::new();
-        claim_collateral_for_self(&mut tx, &objects, &tool_fqn, &owner_cap)
+        claim_collateral_for_self(&mut tx, &objects, &tool, &owner_cap)
             .expect("Failed to build PTB for claiming collateral for a tool.");
         let tx = sui_mocks::mock_finish_transaction(tx);
         let sui::types::TransactionKind::ProgrammableTransaction(
@@ -585,13 +561,13 @@ mod tests {
 
     #[test]
     fn test_set_invocation_cost() {
-        let tool_fqn = fqn!("xyz.dummy.tool@1");
+        let tool = sui_mocks::mock_sui_object_ref();
         let owner_cap = sui_mocks::mock_sui_object_ref();
         let objects = sui_mocks::mock_nexus_objects();
         let invocation_cost = 500;
 
         let mut tx = sui::tx::TransactionBuilder::new();
-        set_invocation_cost(&mut tx, &objects, &tool_fqn, &owner_cap, invocation_cost)
+        set_invocation_cost(&mut tx, &objects, &tool, &owner_cap, invocation_cost)
             .expect("Failed to build PTB for setting invocation cost.");
         let tx = sui_mocks::mock_finish_transaction(tx);
         let sui::types::TransactionKind::ProgrammableTransaction(