+++ CHANGELOG.md
@@ -14,6 +14,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `signed_http` feature and module for application layer HTTP request/response signatures.
 - `network_auth` helpers, types, and PTB templates for tool key registration and leader allowlists.
 
+#### Changed
+
+- adjusted transaction templates and events to support tools as derived objects
+- removed all `Tool*` types for a unified `Tool` type that supports both offchain and onchain tools and `ToolRef` type to differentiate
+
 ### `nexus-cli`
 
 #### Added
@@ -22,6 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `SharedObjectRef` type to represent shared object references with mutability information
 - `AnnounceInterfacePackageEvent` now has `shared_objects` field of type `Vec<SharedObjectRef>` that carries the reference type information
 - added support for tagged_output in sdk
+
+#### Changed
+
+- `nexus tool list` now works with pagination and derived objects
 
 ### `nexus-toolkit-rust`